Move Requirements_Check to a clean class file
